requirements.txt
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
grpcio==1.53.0
grpcio-tools==1.53.0
itsdangerous==2.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
protobuf==4.22.1
pygame==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
